picture + asset refactoring
