+++ src/App.scss
@@ -1,7 +1,7 @@
 // Fonts 
 @font-face {
   font-family: 'Neue Machina';
-  src: url('./fonts/NeueMachina-Variable.ttf') format('truetype');
+  src: url('./assets/fonts/NeueMachina-Variable.ttf') format('truetype');
        font-weight: 100 900;
        font-display: swap;
        font-style: normal;
@@ -9,7 +9,7 @@
 
 @font-face {
   font-family: 'AeonikMono';
-  src: url('./fonts/AeonikMono-Regular.otf') format('opentype');
+  src: url('./assets/fonts/AeonikMono-Regular.otf') format('opentype');
        font-weight: 400;
        font-display: swap;
        font-style: normal;
