+++ src/assets/fonts/AeonikMono-Regular.otf
