+++ src/assets/icons/animatedArrow.js
