+++ src/assets/icons/deepDiveIcon.js
