+++ src/assets/images/logo.js
