+++ src/components/gallery/GalleryCard.tsx
@@ -4,6 +4,9 @@ import React from "react";
 import { CardData } from "./types";
 import { hoverCard } from "./helpers";
 import { useNavigate  } from "react-router-dom";
+import honeybadger from '../../assets/images/honeybadger.jpg';
+import { useLoader } from "@react-three/fiber";
+import { TextureLoader } from "three";
 
 
 interface GalleryCardtProps {
@@ -14,6 +17,9 @@ const GalleryCard = (props: GalleryCardtProps) => {
   const {
     cardData
   } = props;
+
+  // TODO discuss how better to work with pictures
+  const texture = useLoader(TextureLoader, honeybadger)
 
   const [hovered, setHovered] = useState(false);
 
@@ -41,7 +47,7 @@ const GalleryCard = (props: GalleryCardtProps) => {
     <animated.mesh position={position} castShadow onClick={onClick} onPointerOver={(e) => onHover(e, true)} onPointerOut={(e) => onHover(e, false)}>
       <planeGeometry args={[2, 1]} />
       {/* @ts-ignore: https://github.com/pmndrs/react-spring/issues/1515 */}
-      <animated.meshStandardMaterial color="blue" transparent opacity={opacity} attach="material" />
+      <animated.meshBasicMaterial transparent opacity={opacity} attach="material"  map={texture}/>
     </animated.mesh>
     </>
   );
